flac: add initial flac decoder based on github.com/go-audio/wav
Also, based on github.com/azul3d/engine/audio/flac

The underlying FLAC decoder is github.com/mewkiz/flac

There currently exists a few remaining bugs to iron out.
For instance, the last few seconds are cut, probably due to
PCMBuffer not being able to handle when the last buffer is not
full. This will be fixed in the future.